Move OptimizerRuleSessionSettingProvider to rule package
