+++ sql/src/main/resources/META-INF/services/io.crate.metadata.settings.session.SessionSettingProvider
@@ -1,1 +1,1 @@
-io.crate.metadata.settings.session.OptimizerRuleSessionSettingProvider
+io.crate.planner.optimizer.OptimizerRuleSessionSettingProvider
